chore: KazanGameType, 스킬 가중치 자료형 int -> float

+++ Source/Tazan/ResourceManager/KazanGameType.h
@@ -426,7 +426,7 @@ struct FSkillBag
 
 	// <SkillID, Weight>
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	TMap<int, int> SkillID_Weight;
+	TMap<int, float> SkillID_Weight;
 };
 
 USTRUCT(BlueprintType)
